requirements.txt
openai
PyPDF2
langchain
faiss-cpu
python-dotenv
tiktoken
fpdf